Separate HTTP server and client code
With a small amount of commonality, these can be in different
packages.
