+++ http/daemon.go
@@ -40,7 +40,7 @@ var (
 )
 
 func NewDaemon(client *http.Client, t flux.Token, router *mux.Router, endpoint string, p platform.Platform, logger log.Logger) (*Daemon, error) {
-	u, err := makeURL(endpoint, router, "RegisterDaemon")
+	u, err := MakeURL(endpoint, router, "RegisterDaemon")
 	if err != nil {
 		return nil, errors.Wrap(err, "constructing URL")
 	}
